refactor: split mcp/mod.rs into mod.rs (protocol) and tools.rs (handlers)
mod.rs (1017 lines): wire types, tool_list, capabilities, elicitation,
MCP server loop, request router, HTTP proxy.

tools.rs (1563 lines): broadcast helpers, dispatch_tool with all
tool handlers, helper functions.

No functional changes — just a separation of concerns for maintainability.

Made-with: Cursor
